pyproject.toml
[project]
name = "doubleml-rag"
version = "0.1.0"
description = "RAG demo over DoubleML documentation — interview prep project"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "anthropic>=0.100.0",
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.5.9",
    "gitpython>=3.1.50",
    "jupyter>=1.1.1",
    "markdown-it-py>=4.2.0",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pdfplumber>=0.11.9",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
    "rich>=15.0.0",
    "ruamel-yaml>=0.19.1",
    "tabulate>=0.10.0",
    "tiktoken>=0.12.0",
    "voyageai>=0.3.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doubleml_rag"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
